slim down socket service, see front end console to get task id, then make POST call with taht id
