+++ app/src/App.tsx
@@ -1,42 +1,9 @@
 import * as React from 'react'
 import GlobalStyles from './GlobalStyles'
 import Mockup from './Mockup'
-import {Task} from './components/AddTaskForm/AddTaskForm'
-
-import * as socketIo from 'socket.io-client';
-
-const SERVER_URL = process.env.SERVER_URL || 'http://localhost:9000';
-
-interface AppState {
-  triggerStatusChange: boolean
-}
 
 class App extends React.Component {
-  state: AppState = {
-    triggerStatusChange: false
-  }
-  private socket: SocketIOClient.Socket;
-
-  constructor(props: any) {
-    super(props);
-
-    this.socket = socketIo(SERVER_URL);
-
-    this.socket.on('task updated', (task: Task) => {
-      this.setState({
-        triggerStatusChange: true
-      });
-    });
-
-    this.updateTask = this.updateTask.bind(this);
-  }
-
-  updateTask(task: Task) {
-    this.socket.emit('update task', task);
-  }
-
   public render() {
-
     return (
       <>
         <GlobalStyles />
